pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "botiot-ddos-detection"
version = "0.1.0"
description = "DDoS IoT detection using Bot-IoT dataset with XGBoost"
authors = [{name = "IOT Project Team"}]
requires-python = ">=3.10"
license = {text = "MIT"}
readme = "README.md"
dependencies = [
    "pandas>=2.0.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "xgboost>=2.0.0,<3.0.0",
    "joblib>=1.3.0,<2.0.0",
    "pyyaml>=6.0,<7.0",
    "matplotlib>=3.7.0,<4.0.0",
    "seaborn>=0.12.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=23.0.0,<24.0.0",
    "flake8>=6.0.0,<7.0.0",
    "ipykernel>=6.25.0,<7.0.0",
    "jupyter>=1.0.0,<2.0.0",
]
demo = [
    "streamlit>=1.28.0,<2.0.0",
    "plotly>=5.17.0,<6.0.0",
]
all = [
    "botiot-ddos-detection[dev,demo]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.venv
  | venv
  | build
  | dist
)/
'''

